BIG COMMIT : RatingBar fully functionnal !  Changes to be committed: 	modified:   app/src/main/java/bam/bam/bam/dataBDD/UserDAO.java 	modified:   app/src/main/java/bam/bam/bam/dataWS/UserJSONParser.java 	modified:   app/src/main/java/bam/bam/bam/modeles/User.java 	modified:   app/src/main/java/bam/bam/bam/modeles/UserNote.java 	modified:   app/src/main/java/bam/bam/globalDisplay/views/SplashScreen.java

+++ app/src/main/java/bam/bam/bam/modeles/UserNote.java
@@ -17,10 +17,10 @@ public class UserNote {
         this.val = 0;
         this.nbVotes = 0;
     }
-    public UserNote(int note)
+    public UserNote(float note)
     {
         this.val = note;
-        this.note = note;
+        this.note = (int)note;
     }
     public UserNote(float note, int nbVotes)
     {
